app/pyproject.toml
[tool.poetry]
name = "perfectdeal-api"
version = "0.1.0"
description = "API for PerfectDeal"
authors = ["Alexandre <moreaualexandre2002@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
strawberry-graphql = "^0.205.0"
pydantic = "^2.11.3"
graphql-core = "^3.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"